+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonPropertyGrid.cs
@@ -228,17 +228,28 @@ namespace Krypton.Toolkit
         private readonly IntPtr _screenDC;
         private bool _alwaysActive;
         private bool _forcedLayout;
-        private KryptonContextMenuItem _resetMenuItem;
+        private readonly KryptonContextMenuItem _resetMenuItem;
+
+        /// <summary>Occurs before a key is pressed while the control has focus.</summary>
+        [Description(@"Occurs before a key is pressed while the control has focus.")]
+        public new event PreviewKeyDownEventHandler? PreviewKeyDown;
+
+        /// <summary>Occurs when the property sort changes.</summary>
+        [Description(@"Occurs when the property sort changes.")]
+        public event EventHandler? PropertySortChanged;
+
+        /// <summary>Occurs when the property tab changes.</summary>
+        [Description(@"Occurs when the property tab changes.")]
+        public event PropertyTabChangedEventHandler? PropertyTabChanged;
+
+        /// <summary>Occurs when the property value changes.</summary>
+        [Description(@"Occurs when the property value changes.")]
+        public event PropertyValueChangedEventHandler? PropertyValueChanged;
+
 
         #endregion
 
-        #region Events
-
-        // TODO:
-
-        #endregion
-
-        #region Constructor
+        #region Identity
 
         /// <summary>Initializes a new instance of the <see cref="KryptonPropertyGrid" /> class.</summary>
         public KryptonPropertyGrid()
@@ -262,6 +273,11 @@ namespace Krypton.Toolkit
             _propertyGrid.Click += OnPropertyGridClick; // SKC: make sure that the default click is also routed.
             _propertyGrid.GotFocus += OnPropertyGridGotFocus;
             _propertyGrid.LostFocus += OnPropertyGridLostFocus;
+            _propertyGrid.PreviewKeyDown += OnPreviewKeyDown;
+            _propertyGrid.PropertySortChanged += OnPropertySortChanged;
+            _propertyGrid.PropertyTabChanged += OnPropertyTabChanged;
+            //_propertyGrid.PropertyChanging += OnPropertyChanging;
+            _propertyGrid.PropertyValueChanged += OnPropertyValueChanged;
 
             _layoutFill = new ViewLayoutFill(_propertyGrid)
             {
@@ -864,6 +880,14 @@ namespace Krypton.Toolkit
             PerformNeedPaint(false);
             _propertyGrid.Invalidate();
         }
+
+        private void OnPreviewKeyDown(object? sender, PreviewKeyDownEventArgs e) => PreviewKeyDown?.Invoke(sender, e);
+
+        private void OnPropertySortChanged(object? sender, EventArgs e) => PropertySortChanged?.Invoke(sender, e);
+
+        private void OnPropertyTabChanged(object? sender, PropertyTabChangedEventArgs e) => PropertyTabChanged?.Invoke(sender, e);
+
+        private void OnPropertyValueChanged(object? sender, PropertyValueChangedEventArgs e) => PropertyValueChanged?.Invoke(sender, e);
 
         /// <inheritdoc />
         protected override void OnEnabledChanged(EventArgs e)